Towards MIDI:
- Converted vector<Sample*> to BufferList and numerous counts from int to ChanCount (and related changes)
- Added fancy type-generic iterators to BufferList, PortIterator (see IO::collect_input for a good example of the idea - the same code will work to read all input (of various types in a single IO, eg instruments) without modification no matter how many types we add)
- Fixed comparison operator bugs with ChanCount (screwed up metering among other things)
- Moved peak metering into it's own object, and moved most of the pan related code out of IO to panner (still a touch more to be done here for MIDI playback)

Not directly MIDI related fixes for problems in trunk:

- Fixed varispeed gain/pan automation to work properly (was reading the wrong range of automation data, probably causing nasty clicks?)
- Fixed crash on varispeed looping (possibly only a 64-bit problem).  It still doesn't work, but at least it doesn't die

Quite a few things broken, and the new classes are pretty filthy still, but I think the direction is a lot better than all my previous plans...


git-svn-id: svn://localhost/ardour2/branches/midi@795 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/ardour/ardour/curve.h
@@ -52,8 +52,8 @@ class Curve : public AutomationList
 	Curve (const Curve& other);
 	Curve (const Curve& other, double start, double end);
 
-	bool rt_safe_get_vector (double x0, double x1, float *arg, int32_t veclen);
-	void get_vector (double x0, double x1, float *arg, int32_t veclen);
+	bool rt_safe_get_vector (double x0, double x1, float *arg, size_t veclen);
+	void get_vector (double x0, double x1, float *arg, size_t veclen);
 
 	AutomationEventList::iterator closest_control_point_before (double xval);
 	AutomationEventList::iterator closest_control_point_after (double xval);
@@ -72,14 +72,14 @@ class Curve : public AutomationList
 	double unlocked_eval (double where);
 	double multipoint_eval (double x);
 
-	void _get_vector (double x0, double x1, float *arg, int32_t veclen);
+	void _get_vector (double x0, double x1, float *arg, size_t veclen);
 
 };
 
 } // namespace ARDOUR
 
 extern "C" {
-	void curve_get_vector_from_c (void *arg, double, double, float*, int32_t);
+	void curve_get_vector_from_c (void *arg, double, double, float*, size_t);
 }
 
 #endif /* __ardour_curve_h__ */
